fix warning

+++ modules/juce_gui_extra/misc/juce_ColourSelector.cpp
@@ -781,10 +781,10 @@ void ColourSelector::labelTextChanged(Label* label)
 void ColourSelector::changeColour()
 {
     if (sliders[0] != nullptr)
-        setCurrentColour (Colour::fromFloatRGBA (sliders[0]->getValue() / 255.f,
-                                                 sliders[1]->getValue() / 255.f,
-                                                 sliders[2]->getValue() / 255.f,
-                                                 sliders[3]->getValue() / 255.f));
+        setCurrentColour (Colour::fromFloatRGBA ((float)sliders[0]->getValue() / 255.f,
+                                                 (float)sliders[1]->getValue() / 255.f,
+                                                 (float)sliders[2]->getValue() / 255.f,
+                                                 (float)sliders[3]->getValue() / 255.f));
 }
 
 //==============================================================================
